requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
bson==0.5.10
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pip==23.3.1
python-dateutil==2.8.2
python-dotenv==1.0.1
rapidfuzz==3.6.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
thefuzz==0.22.1
urllib3==2.2.0
wheel==0.41.2
yarl==1.9.4
